fix(dashboard): resolve lint errors — unused imports, shutterLevel export
- Remove unused Lightbulb/LightbulbOff imports from EquipmentWidget
- Remove unused familyLabel variable from ZoneWidget
- Move shutterLevel utility from WidgetIcons.tsx to widget-icons.ts
  to fix react-refresh/only-export-components lint error
